pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot with verification, Claude integration, and server design features"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3",
    "anthropic>=0.40",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["bot*"]
